+++ __tests__/html/fluentTheme/withDecorator.js
@@ -1,0 +1,5 @@
+/** @jest-environment ./packages/test/harness/src/host/jest/WebDriverEnvironment.js */
+
+describe('Fluent theme applied', () => {
+  test('with decorators', () => runHTML('fluentTheme/withDecorator'));
+});