+++ packages/api/src/decorator/index.ts
@@ -1,0 +1,4 @@
+export { DecoratorComposer } from './private/DecoratorComposer';
+export { default as ActivityDecorator } from './private/ActivityDecorator';
+export { type DecoratorMiddleware } from './private/createDecoratorComposer';
+export { default as ActivityDecoratorRequest } from './private/activityDecoratorRequest';