+++ packages/api/src/decorator/private/activityDecoratorRequest.ts
@@ -1,0 +1,6 @@
+type ActivityDecoratorRequestType = {
+  from: 'bot' | 'channel' | `user` | undefined;
+  state: 'completion' | 'informative' | undefined;
+};
+
+export default ActivityDecoratorRequestType;